Handling "Unhandled" Error message formats.
Currently if we have any non formatted error message such as without having exception.name , exception.message , exception.stack. We are not tracking the error message instead we end up throwing exception for the track message.

So, changed the code in  a way that if any unhandled error message ( un-formatted ) we are handling in a app inshights code block.